+++ screamingfrog/backends/__init__.py
@@ -3,6 +3,7 @@ from .cli_backend import CLIExportBackend
 from .csv_backend import CSVBackend
 from .db_backend import DatabaseBackend
 from .derby_backend import DerbyBackend
+from .duckdb_backend import DuckDBBackend
 from .hybrid_backend import HybridBackend
 
 __all__ = [
@@ -11,5 +12,6 @@ __all__ = [
     "CSVBackend",
     "DatabaseBackend",
     "DerbyBackend",
+    "DuckDBBackend",
     "HybridBackend",
 ]